Recategorize constants, file reading w.i.p
