--- src/main/java/frc/robot/Constants.java
@@ -1,28 +1,0 @@
-package frc.robot;
-
-public class Constants {
-    public class XBConstants {
-        public static final int drivePort = 0;
-        public static final int opPort = 1;
-    }
-
-    public class DrivetrainConstants {
-        public static final int kLeftFrontPort=0;
-        public static final int kLeftBackPort=1;
-        public static final int kLeftTopPort=2;
-
-        public static final int kRightFrontPort=3;
-        public static final int kRightBackPort=4;
-        public static final int kRightTopPort=5;
-    }
-
-    public class ArmConstants {
-        public static final int kPivotPort1 = 6;        
-        public static final int kPivotPort2 = 7;
-        public static final int kExtensionPort = 8;
-            
-        public static final int kPivotEncoderPort = 0;
-        public static final int kExtensionEncoderSourceA = 0;
-        public static final int kExtensionEncoderSourceB = 0;
-    }
-}
