+++ src/main/java/frc/robot/constants/ArmConstants.java
@@ -1,0 +1,11 @@
+package frc.robot.constants;
+
+public class ArmConstants {
+    public static final int kPivotPort1 = 6;        
+    public static final int kPivotPort2 = 7;
+    public static final int kExtensionPort = 8;
+        
+    public static final int kPivotEncoderPort = 0;
+    public static final int kExtensionEncoderSourceA = 0;
+    public static final int kExtensionEncoderSourceB = 0;
+}
