+++ src/main/java/frc/robot/constants/XBConstants.java
@@ -1,0 +1,7 @@
+package frc.robot.constants;
+
+public class XBConstants {
+    public static final int drivePort = 0;
+    public static final int opPort = 1;
+}
+
